Move Report to utils.
