+++ blackbox/handlers/notifiers/_base.py
@@ -1,23 +1,7 @@
-import dataclasses
 from abc import abstractmethod
 
 from blackbox.handlers._base import BlackboxHandler
-from blackbox.utils.reports import DatabaseReport
-
-
-@dataclasses.dataclass
-class Report:
-    databases: list[DatabaseReport] = dataclasses.field(default_factory=list)
-
-    @property
-    def success(self) -> bool:
-        """Return whether or not the workflow is a success."""
-        return all(report.success for report in self.databases)
-
-    @property
-    def output(self) -> str:
-        """Return the combined outputs from all the database reports."""
-        return "\n".join(report.output for report in self.databases)
+from blackbox.utils import reports
 
 
 class BlackboxNotifier(BlackboxHandler):
@@ -26,9 +10,9 @@ class BlackboxNotifier(BlackboxHandler):
     handler_type = "notifier"
 
     def __init__(self, **kwargs):
-        """Set up notiifer handler."""
+        """Set up notifier handler."""
         super().__init__(**kwargs)
-        self.report = Report()
+        self.report = reports.Report()
 
     @abstractmethod
     def _parse_report(self) -> dict:
@@ -40,6 +24,6 @@ class BlackboxNotifier(BlackboxHandler):
         """Send a notification to the configured notifier."""
         raise NotImplementedError
 
-    def add_database(self, report: DatabaseReport):
+    def add_database(self, report: reports.DatabaseReport):
         """Add a database report to the current list of reports."""
         self.report.databases.append(report)
